+++ test/e2e/scrapeconfig_test.go
@@ -1779,23 +1779,23 @@ var ScrapeConfigCRDTestCases = []scrapeCRDTestCase{
 		expectedError: false,
 	},
 	{
-		name: "ScrapeFallbackProtocol: Valid Value",
+		name: "FallbackScrapeProtocol: Valid Value",
 		scrapeConfigSpec: monitoringv1alpha1.ScrapeConfigSpec{
-			ScrapeFallbackProtocol: ptr.To(monitoringv1.OpenMetricsText0_0_1),
+			FallbackScrapeProtocol: ptr.To(monitoringv1.OpenMetricsText0_0_1),
 		},
 		expectedError: false,
 	},
 	{
-		name: "ScrapeFallbackProtocol: Invalid Protocol",
+		name: "FallbackScrapeProtocol: Invalid Protocol",
 		scrapeConfigSpec: monitoringv1alpha1.ScrapeConfigSpec{
-			ScrapeFallbackProtocol: ptr.To(monitoringv1.ScrapeProtocol("InvalidProtocol")),
+			FallbackScrapeProtocol: ptr.To(monitoringv1.ScrapeProtocol("InvalidProtocol")),
 		},
 		expectedError: true,
 	},
 	{
-		name: "ScrapeFallbackProtocol: Setting nil",
+		name: "FallbackScrapeProtocol: Setting nil",
 		scrapeConfigSpec: monitoringv1alpha1.ScrapeConfigSpec{
-			ScrapeFallbackProtocol: nil,
+			FallbackScrapeProtocol: nil,
 		},
 		expectedError: false,
 	},